rename file form confusing type.d.ts file to types.ts file
